prescription-ocr/requirements.txt
Flask
pillow
pytesseract
opencv-python
flask-cors